Padronizar validações, unificar limpeza de campos
